test(best-practices): downloaded all test cases from their remote servers
Fixtures should be local and always available. Many of the test case files seem to be hosted in user home folders and are probably at risk of being deleted eventually.

The files should definitely be moved to a more streamlined set of folders instead of being based on their original users as in this commit.