wip: clean up of `EvaluationModuleDefinition` and `ACTRules`
There seems to be some issues with the chain of execution from constructor to `.configure()` that is very delicate and breaks easily.

I've rewritten `EvaluationModuleDefinition` to use a generic type parameter for the internal `Tester` object, so inheriting classes can define specializations and have them be visible internally. Further, I've moved several abstract fields into the constructor proper. I'm not sure inheritance was properly forcing those fields to be set, but putting them in the constructor signature forces that case.