test(act-rules): clean up and fix unit tests
Remove global declarations for `window` again. It seems to bleed into other parts of code (probably due to the tsconfig that pulls in unit tests when writing dev code). I don't know what the "most" elegant way of solving it is. @ts-expect-error is a bit verbose but at least describes the issue on the problematic line.